requirements.txt
docker-py==0.2.3
mock==1.0.1
requests==1.2.3
six==1.3.0
websocket-client==0.11.0
click==3.1
